fixed comments in the PR
